change some settings and add nnn cd function

+++ .bashrc
@@ -1,6 +1,6 @@
 case "$-" in
-	*i*) ;;
-	*) return ;;
+*i*) ;;
+*) return ;;
 esac
 
 shopt -s autocd
@@ -13,15 +13,34 @@ alias mv='mv -iv'
 alias rm='rm -vI'
 alias mkdir='mkdir -pv'
 
-alias ll='ls -alF'
-alias la='ls -A'
-alias l='ls -CF'
+alias ll='ls -lAF'
+alias la='ls -AF'
 
 alias ls='ls --color=auto'
 alias grep='grep --color=auto'
 
+alias v='nvim'
+
 skli() {
-    rm config.h
-    make || return
-    doas make install
+	[ -f config.h ] && rm config.h
+	make || return
+	doas make install
 }
+
+ncd() {
+	if [ "${NNNLVL:-0}" -ne 0 ]; then
+		echo "nnn is already running"
+		return
+	fi
+
+	export NNN_TMPFILE="${XDG_CONFIG_HOME:-$HOME/.config}/nnn/.lastd"
+
+	command nnn "$@"
+
+	if [ -f "$NNN_TMPFILE" ]; then
+		. "$NNN_TMPFILE"
+		rm -f -- "$NNN_TMPFILE" >/dev/null
+	fi
+}
+
+bind '"\C-f": "ncd\r"'
